Implement "running" circuit
Running the circuit implies that UI updates are disabled (the
components will no longer emit graphic change signals).
From this, we may "run" the circuit in a separate thread, as to not
block the GUI thread while maximum single-threaded execution speed is
reached.

+++ interface/vsrtl_interface.h
@@ -436,12 +436,12 @@ public:
         return getSubComponents([=](SimComponent& c) { return c.isSynchronous(); });
     }
 
-    unsigned getCycleCount() const { return m_cycleCount; }
+    long long getCycleCount() const { return m_cycleCount; }
 
     virtual void setSynchronousValue(SimSynchronous* c, VSRTL_VT_U addr, VSRTL_VT_U value) = 0;
 
 protected:
-    unsigned m_cycleCount = 0;
+    long long m_cycleCount = 0;
     bool m_emitsSignals = true;
 };
